refactor(llm): split sdk.ts into internal/** modules with data-driven provider capabilities (#414)
Behaviour-preserving restructure of controller/src/llm/. The eight public
modules (sdk/provider/dj/agent/tools/segment-tools/speech/log) are now thin
barrels over llm/internal/**, grouped by concern (core, provider, strategy,
prompts, tools, speech, telemetry).

- The 968-line sdk.ts splits into core/ (pure helpers, retry, failover) and
  strategy/ (djText, djObject, djAgent, object-via-tool).
- Per-provider conditionals collapse into data-driven capability descriptors
  (internal/provider/capabilities.ts): thinking-knob translation + structured-
  output strategy declared once per provider, so strategy code has no
  'provider ===' branches.
- djAgent's native-first -> done-tool -> recovery -> text-salvage cascade
  dispatches off a pure agentPlan() (strategy/plan.ts).
- New pure-helper unit test seam: scripts/llm-pure.test.ts (npm run test:llm),
  pinning the failover gate, per-provider thinking knobs, plan routing, JSON
  salvage, and intro-budget logic.

Public API is byte-identical -- every export keeps its name/signature/return
shape, the ~14 callers are untouched, and the 'via' telemetry taxonomy is
preserved (stats.ts still matches 'ai-sdk:agent').

Verified: npm run lint (0 errors), npm run test:llm (21/21), runtime barrel
load smoke, and live across ollama (glm-5.1:cloud done-tool 16/16), deepseek,
and openrouter (native path) + the picker-benchmark matrix. Targets ai@6.0.206.
