+++ controller/src/llm/internal/prompts/context.ts
@@ -1,0 +1,103 @@
+// The shared "right now" context block + narrative-variety helpers. Used by
+// every generate* script, by matchRequest, and by the segment director
+// (skills/_agent.js) — so they all show the model the same picture of the moment.
+
+// Narrative angles per call type. One is picked at random and injected into the
+// user prompt as "Tone for this segment:" so consecutive generations don't fall
+// back to the same shape. Only the generate* callers consume these — the segment
+// director gets its variety from its CAPABILITIES descriptions. Add freely — the
+// more variety here, the less the DJ repeats itself.
+export const ANGLES = {
+  intro: [
+    'Open with one specific image from right now (weather, time, day, season) and slide into the track.',
+    'Mention the artist in passing — one detail (era, scene, mood) — not a full title-and-artist back-announce.',
+    'Skip the introduction entirely and start mid-thought, as if continuing a conversation.',
+    'React to the request itself — what kind of request it is, what mood it suggests — before mentioning the track.',
+    'Use a short personal observation about the moment (Tuesday energy, the rain holding off, etc.) as the doorway.',
+    'Lean into contrast: how this track sits against what came before, or against the time of day.',
+    'Just say one true sentence and let the music start.',
+  ],
+  link: [
+    'Comment on a contrast or similarity between the two tracks (era, mood, instrumentation, tempo).',
+    'Tie the next track to the time of day, weather, or season — specifically, not generically.',
+    'Mention something small and tactile about right now (the rain, the dark, the smell of coffee, the day of the week).',
+    'Reference the previous artist or song obliquely — one detail, no full back-announce.',
+    'Skip the back-announce entirely and just open a small thought about what is next.',
+    'Acknowledge a listener-shaped moment (commute, late shift, weekend, midweek lull) without naming any listener.',
+    'Make one quiet observation that has nothing to do with either track and let the next song answer it.',
+  ],
+  station_id: [
+    'Plain ident — say the station name and the DJ name, nothing else.',
+    'Anchor the ident to the current moment (a Tuesday afternoon, a foggy evening, the slow part of Sunday).',
+    'Make it a near-aside: like someone reminding themselves where they are.',
+    'Open with the time or weather, then drop the station name in the middle of the sentence.',
+    'A single observation about broadcasting from a homelab, with the station name woven in.',
+  ],
+  hourly: [
+    'State the time as a small fact, then anchor it with one observation about the day.',
+    'Treat the hour mark like a quiet check-in, not a bulletin.',
+    'Open with where in the day we are (mid-afternoon lull, evening getting started, etc.) before the actual time.',
+    'Just one short sentence that happens to mention the time.',
+    'Acknowledge what kind of listener might be tuning in at this exact hour, without naming them.',
+  ],
+};
+
+export function pickAngle(kind: string) {
+  const list = (ANGLES as any)[kind];
+  if (!list || list.length === 0) return null;
+  return list[Math.floor(Math.random() * list.length)];
+}
+
+export function randomSeed() {
+  return Math.floor(Math.random() * 1_000_000_000);
+}
+
+export function buildContextLines(context: any, { recentTracks }: { recentTracks?: any[] } = {}) {
+  const lines: string[] = [];
+  if (context?.date) {
+    lines.push(`Day: ${context.date.dayLabel}, ${context.date.dayOfMonth} ${context.date.monthLabel} (${context.date.season})`);
+  }
+  if (context?.clock) {
+    const tags: string[] = [];
+    if (context.clock.isWeekend) tags.push('weekend');
+    if (context.clock.isLateNight) tags.push('late night');
+    if (context.clock.isCommute) tags.push('commute hour');
+    lines.push(`Local time: ${context.clock.hhmm}${tags.length ? ' · ' + tags.join(' · ') : ''}`);
+  }
+  if (context?.time) lines.push(`Period: ${context.time.period} (${context.time.vibe})`);
+  if (context?.weather && context.weather.condition && context.weather.condition !== 'unknown') {
+    lines.push(`Weather in ${context.weather.location}: ${context.weather.condition}${context.weather.temp != null ? `, ${context.weather.temp}°${context.weather.tempUnit || 'C'}` : ''}`);
+  }
+  if (context?.festival) lines.push(`Festival: ${context.festival.name}`);
+  if (context?.activeShow) {
+    const topic = context.activeShow.topic ? ` — ${context.activeShow.topic}` : '';
+    lines.push(`On now: the show "${context.activeShow.name}"${topic}. Stay loosely on its theme.`);
+  }
+  if (context?.listeners?.count != null) {
+    const n = context.listeners.count;
+    lines.push(n === 0
+      ? `No one is tuned in right now.`
+      : `Listeners tuned in right now: ${n}.`);
+  }
+  if (recentTracks && recentTracks.length) {
+    const list = recentTracks.slice(0, 5).map((t: any) => `"${t.title}" by ${t.artist || 'unknown'}`).join('; ');
+    lines.push(`Recently played (do not mention these artists or titles): ${list}`);
+  }
+  return lines;
+}
+
+// Append rotating angle + recap + opener blocklist to the user prompt.
+export function decoratePrompt(
+  prompt: string,
+  { kind, recap, recentOpeners }: { kind: string; recap?: string | null; recentOpeners?: string[] | null },
+) {
+  const out: string[] = [prompt];
+  const angle = pickAngle(kind);
+  if (angle) out.push(`\nTone for this segment: ${angle}`);
+  if (recap) out.push(`\nYou said these things on-air recently (do not repeat phrasing or topics):\n${recap}`);
+  if (recentOpeners && recentOpeners.length) {
+    const list = recentOpeners.slice(0, 6).map((o: string) => `"${o}…"`).join(', ');
+    out.push(`\nDo not start your line with any of these openers (vary the first words): ${list}`);
+  }
+  return out.join('\n');
+}
