+++ controller/src/llm/internal/prompts/picker.ts
@@ -1,0 +1,72 @@
+// LLM pool picker — choose the next track from a candidate pool (the stateless
+// fallback path; the conversational agent picker lives in broadcast/dj-agent.js).
+// PICKER_CRITERIA is shared with that agent so the two strategies can't drift.
+
+import { z } from 'zod';
+import * as settings from '../../../settings.js';
+import { djObject } from '../strategy/object.js';
+
+export const PICKER_CRITERIA = `Selection criteria, in order:
+1. FLOW — does it transition naturally from what just played (energy, mood, tempo)? When a candidate shows a "bpm" and/or Camelot "key", those are MEASURED — prefer a next track whose tempo sits near the current one (or steps it deliberately for the daypart) and whose key is harmonically close. A "pace" value (0–1) is the track's MEASURED perceptual energy, decoupled from tempo — use it to shape build/release arcs: avoid stacking two peaks back-to-back, ease down for wind-down dayparts, lift for workout/drive. A "sections" count hints how much the opening develops (higher = a busier, evolving intro). Treat all of these as tie-breakers, never hard rules; many tracks won't have them.
+2. CONTEXT — does it fit the time of day, weather, and dominant mood?
+3. VARIETY — avoid the same artist back-to-back; don't repeat tracks you've already played today; rotate energy. Variety over cleverness — never pick a track because its title literally matches the time of day, the weather, or anything else literal.
+4. INTEREST — prefer something that creates a moment, not the most generic option.`;
+
+function pickerSystem(show?: { name: string; topic: string } | null) {
+  const stationName = settings.get().station;
+  const showLine = show?.topic
+    ? `\n\nCurrent show brief — follow this for every pick:\n${show.topic}`
+    : '';
+  return `You are the DJ for ${stationName}, a personal internet radio station.
+Pick the single best NEXT track from the candidate pool, given recent plays and the current context.${showLine}
+
+${PICKER_CRITERIA}
+
+Each candidate carries a "source" tag — a hint about where it came from:
+- similar / similar-artist: flows from what's playing now
+- embedding-similar: closest in mood / lyric / metadata space to what's playing
+- audio-similar: SOUNDS closest to what's playing (timbre, instrumentation, production)
+- audio-journey: SOUNDS like where the set is heading — the next step of a deliberate drift toward a destination vibe, not necessarily the current track
+- recent: newly added to the library
+- frequent / starred / playlist: an established favourite
+- mood-library: matches the room's mood
+- random: a wildcard for breaking a predictable run
+Use it to balance familiarity against discovery. The two *-similar sources may
+carry a "similarity" (0–1, higher = closer) — a high value means a very tight
+match you can lean on for a smooth segue.
+
+recentPlays is context for judging flow — every candidate is already guaranteed
+unplayed, so you never need to reject one for being recent.
+
+Pick exactly one candidate.`;
+}
+
+export async function pickNextTrack({ candidates, recentPlays, context, show = null }: {
+  candidates: any[];
+  recentPlays: any;
+  context: any;
+  show?: { name: string; topic: string } | null;
+}) {
+  const user = JSON.stringify({
+    now: {
+      time: context.time?.period,
+      vibe: context.time?.vibe,
+      mood: context.dominantMood,
+      weather: context.weather?.condition,
+      festival: context.festival?.name,
+    },
+    recentPlays,
+    candidates,
+  }, null, 2);
+
+  return djObject({
+    system: pickerSystem(show),
+    prompt: user,
+    schema: z.object({
+      id: z.string().describe('the exact id of one candidate'),
+      reason: z.string().describe('one short sentence on why this one'),
+    }),
+    temperature: 0.5,
+    kind: 'pickNextTrack',
+  });
+}
