+++ controller/src/llm/internal/prompts/system.ts
@@ -1,0 +1,56 @@
+// DJ system prompt + persona-driven verbosity. Resolves the prompt for the
+// persona on air right now (the current show's owner if one is scheduled,
+// otherwise the admin-selected active persona — settings.getEffectivePersona).
+
+import * as settings from '../../../settings.js';
+
+// Paralinguistic tags Chatterbox renders as actual non-verbal sounds. Every
+// other engine (piper, kokoro, cloud) reads `[laugh]` aloud as the word
+// "laugh", so we only mention this when the on-air persona will actually be
+// voiced by Chatterbox.
+const CHATTERBOX_TAG_HINT =
+  '\n\nYou may sparingly insert non-verbal cues in square brackets: [laugh], [chuckle], [sigh], [cough]. Use them only where genuinely natural — at most one per segment, and never as filler.';
+
+export function djSystem() {
+  const persona = settings.getEffectivePersona();
+  const s = settings.get();
+  const base = settings.renderDjPrompt(persona, {
+    station: s.station,
+    location: s.weather?.locationName,
+  });
+  if (persona?.tts?.engine === 'chatterbox') return base + CHATTERBOX_TAG_HINT;
+  return base;
+}
+
+// Persona-driven verbosity. 'concise' reproduces the historical one-liner
+// segment lengths; 'extended' roughly doubles every segment so a storytelling
+// persona can stretch out. Resolved from the on-air persona, the same way
+// djSystem() resolves it — see settings.getEffectivePersona / SCRIPT_LENGTHS.
+const LENGTH_PHRASES = {
+  concise: {
+    intro:     'Keep it brief — 2 to 4 sentences.',
+    link:      '1-2 sentences',
+    stationId: 'a 1-sentence station ident',
+    hourly:    '1 sentence',
+    adlib:     '1-2 sentences',
+    segment:   'one sentence',
+  },
+  extended: {
+    intro:     'Take your time — 5 to 8 sentences. Set a scene, tell a small story around the track.',
+    link:      '4-6 sentences',
+    stationId: 'a 2-3 sentence station ident',
+    hourly:    '2-3 sentences',
+    adlib:     '4-6 sentences',
+    segment:   'three to five sentences',
+  },
+};
+
+export function lengthMode(persona: any = settings.getEffectivePersona()) {
+  return persona?.scriptLength === 'extended' ? 'extended' : 'concise';
+}
+
+// The length directive for one segment kind, for the on-air (or given) persona.
+export function lengthPhrase(kind: string, persona?: any) {
+  const m = (LENGTH_PHRASES as any)[lengthMode(persona)];
+  return m[kind] || m.link;
+}
